\name{distance2}
\alias{distance2}

\title{
Euclidean Distance between two matrices
}
\description{
Euclidean Distance between two matrices
}
\usage{
distance2(x1, x2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x1}{
Matrix 1
}
  \item{x2}{
Matrix 2
}
}
 
\author{
Suhai (Timothy) Liu
}


\examples{
X<-matrix(1:12,3,4)
Y<-matrix(5:16,3,4)
distance2(X,Y)

} 
